Combine shape point and marcherCurve
